+++ ohsome_quality_api/oqt.py
@@ -18,7 +18,6 @@ async def create_indicator(
     bpolys: FeatureCollection,
     topic: TopicData | TopicDefinition,
     include_figure: bool = True,
-    *args,
     **kwargs,
 ) -> list[Indicator]:
     """Create indicator(s) for features of a GeoJSON FeatureCollection.
@@ -30,14 +29,15 @@ async def create_indicator(
     for i, feature in enumerate(bpolys.features):
         if "id" not in feature.keys():
             feature["id"] = i
-        # Only enforce size limit if ohsome API data is not provided
         # Disable size limit for the Mapping Saturation indicator
+        # TODO: Remove size restriction
         if isinstance(topic, TopicDefinition) and key not in [
             "mapping-saturation",
             "currentness",
             "building-comparison",
             "road-comparison",
             "attribute-completeness",
+            "land-cover-thematic-accuracy",
         ]:
             validate_area(feature)
         tasks.append(
@@ -46,7 +46,6 @@ async def create_indicator(
                 feature,
                 topic,
                 include_figure,
-                *args,
                 **kwargs,
             )
         )
@@ -58,7 +57,6 @@ async def _create_indicator(
     feature: Feature,
     topic: Topic,
     include_figure: bool = True,
-    *args,
     **kwargs,
 ) -> Indicator:
     """Create an indicator from scratch."""
@@ -71,7 +69,6 @@ async def _create_indicator(
     indicator = indicator_class(
         topic,
         feature,
-        *args,
         **kwargs,
     )