+++ tests/integrationtests/utils.py
@@ -1,11 +1,9 @@
 import inspect
 import os
-from pathlib import Path
 from unittest.mock import MagicMock
 
 import geojson
 import vcr
-from approvaltests.namer.namer_base import NamerBase
 
 from ohsome_quality_api.topics.definitions import get_topic_preset
 from ohsome_quality_api.topics.models import TopicDefinition
@@ -13,41 +11,6 @@ from ohsome_quality_api.topics.models import TopicDefinition
 TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 FIXTURE_DIR = os.path.join(TEST_DIR, "fixtures")
 VCR_DIR = os.path.join(FIXTURE_DIR, "vcr_cassettes")
-APPROVED_DIR = os.path.join(TEST_DIR, "approved")
-
-
-class PytestNamer(NamerBase):
-    def __init__(self, extension=None):
-        """An approval tests Namer for naming approved and received text files.
-
-        These files will get stored under:
-        `tests/approval/{module}/test_file.py--TestClass--test_func[a]`
-
-        This class utilizes the `PYTEST_CURRENT_TEST` environment variable, which
-        consist of the nodeid and the current stage:
-        `relative/path/to/test_file.py::TestClass::test_func[a] (call)`
-
-        During a pytest test session stages can be setup, teardown or call.
-        Approval tests should only be used during the call stage and therefore
-        the ` (call)` postfix is removed.
-
-        To avoid forbidden characters in system paths `::` is replaced by `-`.
-        """
-        self.nodeid: Path = Path(os.environ["PYTEST_CURRENT_TEST"])
-        NamerBase.__init__(self, extension)
-
-    def get_file_name(self) -> Path:
-        """File name is pytest nodeid w/out directory name and pytest stage."""
-        return Path(str(self.nodeid.name).replace(" (call)", "").replace("::", "-"))
-
-    def get_directory(self) -> Path:
-        """Directory is `tests/approval/{module}`."""
-        parts = self.nodeid.parent.parts
-        raw = Path(*[p for p in parts if p not in ["tests", "integrationtests"]])
-        return Path(APPROVED_DIR) / raw
-
-    def get_config(self) -> dict:
-        return {}
 
 
 class AsyncMock(MagicMock):
@@ -113,6 +76,7 @@ oqapi_vcr = vcr.VCR(
     match_on=["method", "scheme", "host", "port", "path", "query", "body"],
     func_path_generator=filename_generator,
     before_record_response=replace_body(["image/png"], dummy_png),
-    ignore_hosts=["testserver"],
+    # ignore github.com. ApprovalTests
+    ignore_hosts=["testserver", "github"],
     ignore_localhost=True,  # do not record HTTP requests to local FastAPI test instance
 )